Rename enum module to plugins_enum
