+++ spyder/api/plugins_enum.py
